breaking(transport): Make HTTP2Transport the default (#4492)
Making the HTTP2 transport the default when `h2` and `httpcore` packages
are installed. We've been testing this on Sentry SaaS for a while
without any issues.

We should promote installing the SDK as `sentry-sdk[http2]` for this to
be picked up though. Since we still have to support Python 3.7 and `h2`
not being supported there, we cannot install it by default and use HTTP2
directly.

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>