feat: Add LangGraph integration (#4727)
- Add LangGraph integration
- Compilation of StateGraphs results in an agent creation span
(according to [OTEL semantic
conventions](https://opentelemetry.io/docs/specs/semconv/gen-ai/gen-ai-agent-spans/))
- Runtime executions are done on Pregel instances - we are wrapping
their invoke & ainvoke which produces the invoke_agent spans
- There's some internals that automatically switch between invoke &
stream on CompiledStateGraph (which is a subclass of Pregel), which
results in duplicate spans if both are instrumented. For now, only
invoke is wrapped to prevent this duplication.
- Agent handoffs in LangGraph are done via tools - so there is no real
possibility to create handoff spans within the SDK. Looks like this will
be handled in product logic instead.


Closes TET-991
Closes PY-1799

---------

Co-authored-by: Anton Pirker <anton.pirker@sentry.io>