feat(integrations): Support gql 4.0-style execute (#4779)
gql 4.0 [changed](https://github.com/graphql-python/gql/pull/556/files)
the signature of the `execute` function which we were patching. Instead
of a `DocumentNode` it now gets a `GraphQLRequest` (which contains the
`document` attribute with the `DocumentNode`). This means we need to
update the way we're extracting additional data in the event processor.