fix(Django): Avoid evaluating complex Django object in span.data/span.attributes (#4804)
### Description
When rendering templates of the Django Admin in Django 5.0+ the template
context includes also a `QuerySet` with lots of log items.
In our `serialize()` function this was not properly serialized leading
to evaluating the `QuerySet` (read: running the SQL query).

This change updates the `serialize()` to also correctly serialize
`QuerySet`s in `span.data`

#### Issues
* resolves: #4604
* resolves: PY-1781