tests: Move sanic under toxgen & add option to control how many releases to pick (#4767)
- Move sanic under toxgen
- This gets rid of the `-latest` group for Web 2 as all Web 2 test
suites are now governed by toxgen
- Add `num_versions` as a new option to optionally slim down the number
of tested versions (useful for sanic as it has a lot of "majors" due to
calver)
- When a package had multiple majors, we used to always pick the lowest
as well as the highest release in the last major -- removed the lowest
now to not over-test. All majors will still be tested by default, we
just won't be testing two releases in the last major.

Ref https://github.com/getsentry/sentry-python/issues/4506