feat(toxgen): Remove timestamp prone to merge conflicts (#4860)
### Description
Most of the merge conflicts concerning `tox.ini` come from the timestamp
of the last update. If two PRs regenerate `tox.ini`, they'll be
guaranteed to have a merge conflict that needs to be resolved by hand.
This is time consuming and unnecessary.

We still need some way of figuring out when the file was last updated.
This is important for checking unwanted edits to `tox.ini` (instead of
to the template it's generated from). In order to check for them, we
need to regenerate the file on the current branch and compare it to the
committed version. However, if there have been new package versions in
the meantime, these would then pop up in the diff. One solution is to
regenerate the file ignoring any new releases after the last update,
which is what we were using the last updated timestamp for.

In this PR, we simply figure out the timestamp from `git` history,
removing the need for storing it explicitly anywhere.

#### Issues
<!--
* resolves: #1234
* resolves: LIN-1234
-->

#### Reminders
- Please add tests to validate your changes, and lint your code using
`tox -e linters`.
- Add GH Issue ID _&_ Linear ID (if applicable)
- PR title should use [conventional
commit](https://develop.sentry.dev/engineering-practices/commit-messages/#type)
style (`feat:`, `fix:`, `ref:`, `meta:`)
- For external contributors:
[CONTRIBUTING.md](https://github.com/getsentry/sentry-python/blob/master/CONTRIBUTING.md),
[Sentry SDK development docs](https://develop.sentry.dev/sdk/), [Discord
community](https://discord.gg/Ww9hbqr)