ci: Replace `black` and `flake8` with `ruff`. (#4866)
### Description
Replace `black` and `flake8` with `ruff`.

With this change also reformat the code base (only minor changes,
because the configuration is basically the same.) The commits for the
reformatting are not included in `git blame` calls (see
`.git-blame-ignore-revs`).

Also updates the `pre-commit` config so we have exactly the same
linting/formatting in CI and local dev.

We now have:
- Less dependencies 
- Less configuration
- Faster


#### Issues
* resolves: #2527