fix(litestar): Copy request info to prevent cookies mutation (#4883)
Prevent mutating cookies on incoming HTTP requests if the cookie name is
in the scrubbers denylist.

Cookies like `token=...` were replaced with `AnnotatedValue` because a
shallow reference of the request information was held by the client. A
deep copy is introduced so scrubbing does not interfere with Litestar,
and in particular does not break `JWTCookieAuth`.

Closes https://github.com/getsentry/sentry-python/issues/4882

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>