fix(aws): Inject scopes in TimeoutThread exception with AWS lambda (#4914)
Restore ServerlessTimeoutWarning isolation and current scope handling, so the scopes are active in an AWS lambda function and breadcrumbs and tags are applied on timeout. The behavior is restored to that before 2d392af3ea6da91ddbdde55d18e15c24dce6b59b.

More information about how I found the bug is described in
https://github.com/getsentry/sentry-python/issues/4912.

Closes https://github.com/getsentry/sentry-python/issues/4894.