fix(tracemetrics): Bump metric buffer size to 1k (#5031)
We've been noticing some discards from clients for likely buffer reason. Metrics were
initially set to 100 to match logs but the size limit for a metric in relay is ~1000x less (~1-2kb at the moment) so the memory pressure should be fine (it'd cap in the single digit megabyte range).