ci: Run integration tests on Python 3.14t (#4995)
Modifies `populate_tox.py` to add free-threading Python to the test
suite for integrations where we can expect the package to support
free-threading. Packages are considered compatible with free-threading
if the package itself, and all its dependencies, support free-threading.

Pure Python packages are considered compatible if their dependencies
support free-threading. The support for free-threading of packages with
extension modules is determined based on the ABI tag in the wheel name.

A dry-run of `pip install` determines the ABI tags of dependencies.
Relevant output from the command is cached analogously to responses of
PyPI REST calls.

Closes https://github.com/getsentry/sentry-python/issues/4971