feat: Preserve metadata on wrapped coroutines (#5105)
Copy metadata when wrapping coroutines in the patched `asyncio` task
factory. Uses the `functools.update_wrapper()` function that is also
used internally by `functools.wraps()`. Unlike when wrapping functions,
less metadata is available on coroutines, so `_wrap_coroutine()` is
equivalent to `functools.wraps()` but copies fewer attributes.

As `functools.update_wrapper()` checks for the existence of properties
first, copying metadata will not result in an `AttributeError` if
metadata is not available. See
https://docs.python.org/3/library/functools.html#functools.update_wrapper.

The SDK no longer overwrites coroutine metadata included in destroyed
errors as reported in
https://github.com/getsentry/sentry-python/issues/5072.