Add back span status (#5147)
### Description
    
For some reason since Performance was implemented ages ago, we were not
sending the `span.status` field at all.
Add it back since we are moving to a Span-First world and Agent
Monitoring also expects span statuses to be correct for some product
features.
    
Some other notes regarding the status mess:
* the transaction status is added to the `trace_context` - this is fine
* the span status is also added as a tag, this should be removed in a
major since it's redundant
* `set_http_status` further adds another tag called `http.status_code`,
remove this in the major too

#### Issues

* resolves: #5059
* resolves: PY-1960