feat(logs): Record discarded log bytes (#5144)
Report the size of discarded logs in bytes as part of client reports, based on
the specification added in commit 5819cdd of `sentry-docs`.

The reported size is based on the log format submitted to the transport. The
bytes representation returned by `PayloadRef.get_bytes()` is used to calculate
the size in bytes. In all common code paths, the representation is
a compact utf-8 encoding of the log item payload.