feat: Remove string truncation for events (#5095)
Remove the default string limit applied when serializing envelopes.

The `max_value_length` parameter now accepts `None`, which disables string truncation. The default value is changed to `None`.

The change follows on from raising the limit in #4632.

Replaces `DEFAULT_MAX_VALUE_LENGTH` with the value 1024 in tests, which was the original limit.

Closes https://github.com/getsentry/sentry-python/issues/4988