Make PropagationContext.from_incoming_data always return a PropagationContext
When there is any sort of incoming data, and since `continue_trace` is
always intended to be at a system boundary, we always want to force a
new trace if there's no incoming propagation or an incorrect propagation
header.

What previously happened is in these cases, a single `trace_id` was kept
alive in the `propagation_context` even when these were just meant to be
new independent traces (see screenshot).

I think this will actually solve many complaints about long living
traces that were never supposed to be such.