test: Import integrations with empty shadow modules (#5150)
Add a test parametrized on our integrations. The test detects if the integration
imports modules not in the standard library, and if so, verifies that importing
the integration with an empty shadow module raises a `DidNotEnable` exception.

Adding integrations to be auto-enabling can cause SDK crashes in two specific
cases we encountered at the end of last week:
- the user has an old package version that we don’t support, resulting in
an `ImportError` when we patch something that does not yet exist; or
- something like an `agents.py` in the environment shadows the import
and causes an `ImportError` if the auto-activation still triggers.

All integrations with poorly gated imports are not auto-enabling, but some
affected integrations are new (e.g., litellm).

Closes https://github.com/getsentry/sentry-python/issues/5140