ref: Cleanup outgoing propagation_context logic (#5215)
### Description

* Add `get_baggage`, `get_traceparent` and `iter_headers` to
`PropagationContext` so that the contract is similar to `Span`.
* This clarifies that the `PropagationContext` and `Span` are themselves
responsible for constructing these values and not the `Scope`.
* Currently, it is all a mess on `Scope` which does not clarify state
responsibilities correctly on who holds and can populate what.
* Add `Baggage.populate_from_propagation_context` as a replacement for
`Baggage.from_options` (that took a `scope` argument).
* Unify all the various `propagation_context` presence checks in `Scope`
methods to `get_active_propagation_context`.
* This `get_active_propagation_context` logic then remains the central
problem to fix in the future.


### Deprecations

The following methods are thus no longer used internally and deprecated.
* `Scope.get_dynamic_sampling_context`
* `Scope.iter_headers`
* `Baggage.from_options`